finish migration v0
